+++ .codemod/classReplacements.js
@@ -1,3 +1,6 @@
-const classMap = {};
+const classMap = {
+  'xui-fileuploader--fileitem--loading':
+    'xui-fileuploader--fileitem--loading xui-fileuploader--fileitem--loading-indeterminate',
+};
 
 module.exports = classMap;